Fix: testando correções

+++ src/main/java/com/anna/controller/PaymentController.java
@@ -15,9 +15,9 @@ public class PaymentController {
 
     // Processar um pagamento para um pedido
     public void processPayment(int paymentId, Order order, Date date) {
-        double amount = order.getTotalAmount();
+        double amount = order.getTotalAmount().doubleValue(); // Assumindo que getTotalAmount() retorna BigDecimal
         Payment payment = new Payment(paymentId, amount, date);
-        boolean success = payment.processPayment();
+        boolean success = payment.processPayment(); // Aqui precisa ajustar para um método correto de processamento
         if (success) {
             payments.add(payment);
             System.out.println("Pagamento processado com sucesso para o pedido ID: " + order.getOrderId());
